fix method increment input parameter to uint instead of ulong

+++ DAL/NotificationDal.cs
@@ -2483,7 +2483,7 @@ namespace DAL
                     counter = cbManager.Increment(GetUserPushKey(partnerId, userId), 1);
                 else
                 {
-                    ulong docTTL = TCMClient.Settings.Instance.GetValue<ulong>("push_message.ttl_seconds");
+                    uint docTTL = TCMClient.Settings.Instance.GetValue<uint>("push_message.ttl_seconds");
                     if (docTTL == 0)
                         docTTL = TTL_USER_PUSH_COUNTER_DOCUMENT_SECONDS;
